// interfaces/HttpRequestMethod.ts
enum HttpRequestMethod {
  GET = 'get',
  POST = 'post',
  PUT = 'put',
  PATCH = 'patch',
  DELETE = 'delete',
  COPY = 'copy',
  HEAD = 'head',
  OPTIONS = 'options',
  LINK = 'link',
  UNKLINK = 'unlink',
  PURGE = 'purge',
  LOCK = 'lock',
  PROPFIND = 'profind',
  UNLOCK = 'unlock',
  VIEW = 'view',
}
export default HttpRequestMethod;